refactor: move logger options configuration to logger.lib.ts and update request logger middleware for improved logging setup
